[공통 #83] WeQuizDialog util 추가
