--- app/src/main/kotlin/team/ommaya/wequiz/android/base/BaseViewBindingDialogFragment.kt
@@ -1,39 +1,0 @@
-/*
- * Designed and developed by "옴마야" Team 2023.
- *
- * Licensed under the MIT.
- * Please see full license: https://github.com/mash-up-kr/WeQuiz-Android/blob/main/LICENSE
- */
-
-package team.ommaya.wequiz.android.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.appcompat.app.AppCompatDialogFragment
-import androidx.viewbinding.ViewBinding
-
-abstract class BaseViewBindingDialogFragment<VB : ViewBinding>(
-    private val bindingInflater: (inflater: LayoutInflater) -> VB,
-) : AppCompatDialogFragment() {
-
-    private var _binding: VB? = null
-
-    protected val binding: VB
-        get() = _binding ?: error("Binding is null")
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View? {
-        _binding = bindingInflater(inflater)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
